+++ apps/web/app/(marketing)/home/_components/statistics.tsx
