+++ apps/web/app/(marketing)/layout.tsx
@@ -1,7 +1,7 @@
 import { Footer } from '@/app/_components/footer'
 import { Header } from '@/app/_components/header-marketing'
 
-export default function MarkettingLayout({ children }: LayoutProps<'/'>) {
+export default function MarketingLayout({ children }: LayoutProps<'/'>) {
   return (
     <>
       <Header />
